+++ app/src/main/java/com/example/secureyourcourt/utils/FormaUtils.kt
@@ -4,7 +4,7 @@ package com.example.secureyourcourt.utils
 import java.text.NumberFormat
 import java.util.Locale
 
-fun formatPrecio(precio: Double): String {
+fun formatPrecio(precio: Int): String {
     val formato = NumberFormat.getNumberInstance(Locale("es", "CL"))
     return "$" + formato.format(precio.toInt())
 }
